% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{heatmap_calendar}
\alias{heatmap_calendar}
\alias{heatmap_calendar-package}
\title{heatmap_calendar}
\usage{
heatmap_calendar(mydata)
}
\arguments{
\item{file}{read in a csv file.}
}
\description{
heatmap_calendar

This function creates one heatmap calendar.
}
\examples{
heatmap_calendar()
}
\keyword{data}
\keyword{event}

