% Generated by roxygen2 (4.0.1): do not edit by hand
\name{heatmap_calendar_big}
\alias{heatmap_calendar_big}
\title{A function to create heatmap calendars from event data.}
\usage{
heatmap_calendar_big(mydata)
}
\arguments{
\item{file}{read in a csv file.}
}
\description{
This function creates four heatmap calendars from an uploaded csv file.
}
\examples{
heatmap_calendar()
}
\keyword{data}
\keyword{event}

