% Generated by roxygen2 (4.0.1): do not edit by hand
\name{heatmap_calendar_big}
\alias{heatmap_calendar_big}
\title{A function to create heatmap calendars from event data.}
\usage{
heatmap_calendar_big(mydata)
}
\arguments{
\item{file}{read in a csv file.}
}
\description{
This function creates one heatmap calendar with a bar chart above.
}
\examples{
heatmap_calendar_big()
}
\keyword{data}
\keyword{event}

